Convert module from MakeMaker -> Module::Build
I am not really sure what was going on before (The module was
previously maintained by Marcel Gruenauer) and it looks like
he was using dist-zilla, but with a bunch of his personal
configs, however there also looks like a lot of MakeMaker as
well.

This has made the some of the tests  slightly unworkable, so
they have been removed.  Never mind, it's Module::Build for now.