taint mode not respected
If the -T argument is passed on the command line to the perl executable,
it will turn on "taint" treating all input as suspect until checked,
and dieing if the scripts attempts to output tainted data. Using taint
mode is considered good practice for sensitive programs that could
possibly be run by untrusted users. If the -T argument is used in the
shebang line of the script, then it needs to be passed when the script is
invoked - otherwise the script will fail to compile. Thus it is quite
important that this module pass the -T flag when required. We also provide
a test script to verify the extra functionality. There is also a -t argument
which warns rather than dies.

Last-Update: 2012-02-18
Bug-Debian: http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=649301
Bug: https://rt.cpan.org/Public/Bug/Display.html?id=55837