need to provide path to libraries of distribution
Imagine you have a distrubution consisting of a script and at least
one module. You want to test that the script compiles using this
module. However older versions of your module are already installed.
It is important that we pull in the current version of the modules
not the already released versions. The upstream code does not
take care of this at all. This patch also provides a test.

Bug: http://rt.cpan.org/Ticket/Display.html?id=72557
Bug-Debian: http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=649332
Last-Update: 2012-02-18